Add message templates for sending institutional messages programmatically
Message templates are localizable and use pebble as a templating language.
LocalizedString arguments are localized automatically when sending the
message. The templates are defined via annotation to be registered on
servlet container initialization. This annotation allows initializing the
template content through the .properties files and the messaging
configuration interface allows editing the templates' contents in
production.
The institutional footer with recipient descrimination was converted to a
template that wraps any message sent via the interface and can now be
edited in production.
Minor changes and corrections were made to message sending, mostly related
to exception throwing, and to the messaging interfaces.