Allow senders to send messages to ad hoc recipients Allow locking message interface to a specific sender
Recipients to be used in the message sending interface are now signed by the server using JWT.
Through these server provided recipient tokens it is now possile for a sender to send a message
to any recipient group. Use MessageBean::addAdHocRecipient to provide these tokens and
MessageBean::selectRecipient to pre-select a recipient in the interface. Direct usage of the
bean's setters is discouraged.
To prevent the current user of switching senders in situations where it does not make sense,
such as in the workflow of a role-specific sender task, it is now possible to lock the
interface to a pre-selected sender by using MessageBean::setLockedSender.